Supported all nested stack configurations
This started with trying to get the medley sample to work. Navigate forward, open modal, navigate forward, close modal and the predictive back did nothing. The modal still renders the nested modal stack, it still has a back stack and is the primary fragment so it receives the back even though it’s closed. (The predictive back doesn’t work in modal but that’s a different issue - android modal receives the back even though the stack listener is registered after).

Original idea was to shift the primary stack to the main stack when the modal stack is detached from the window. This worked but wouldn’t work for a stack directly inside a stack (without a modal). Navigate A → B in main stack then C → D in stack inside scene B. Predictive back goes back to C and the next predictive back should go back to A but it did nothing. The nested stack is still marked as primary so the main stack won’t receive the back.

The trick is to have multiple primary stacks. Opening a stack inside another stack then both should be primary. Then when the nested stack has used up all the predictive backs to get back to its first scene, then the parent stack will take over the next predictive backs. To do this create the nested stack fragment inside the childFragmentManager of the parent stack fragment. The useId hook gives the perfect way to generate a unique tag for each stack fragment. Storing the tag in React Context means the native side can find the parent fragment.

With sibling stacks, like with stacks inside tabs, then these are all created in the same fragment manager. So only one of them can be primary at any one time. The primary one is the one that’s attached to the window. Tried commitAllowingStateLoss in onAttachedToWindow but this threw an error saying "FragmentManager is already executing transactions". The fragment’s childFragmentManager was already adding scene(s) for the navigation and so couldn’t also add a fragment at the same time. So changed to commit but kept commitAllowingStateLoss for the onDetachedFromWindow (can’t make them both commit because the detach would clear the primary after the attach had already changed it).