Supported React 17
If there's no useId (hook in React 18) then created all stacks in the top level active fragment manager. This works for most scenarios - single stack and tabs with nested stacks. It won't work with stack inside modal because the modal will stay as primary fragment. But this is ok, can updgrade react native to fix