Fixed cancelling gesture error in nested stacks
Take the Twitter sample where Tabs has the ios setup on android with a stack per tab. Set
trackCrumbTrail true on tabs and navigate in home tab Home → Tweet → Tabs. So get 2 rows of bottom tabs. Then starting the predictive back and cancelling it would give IllegalStateException because "already executing transactions".

The onDetachedFromWindow fired on the nested stack but during the parent stack navigation. The predictive back was happening on the parent stack because the nested stack had no back entries. Can’t set the primary fragment when the fragment manager is changing the back stack.

Changed to catch the exception. Mustn’t set the primary stack to null in this case because it still is the primary stack. If later navigate forward in this nested stack then it must receive the next predictive back and so must be primary.