Changed to commit now for safety
Better to commit now because almost all the time it will work and is safer. Want to change the primary fragment immediately - if it's delayed then predictive back may notify the wrong fragment manager.

Think the only time it will throw an error is when first mounting the nested stack. Because the parent stack is changing scenes so can't commit until this is finished. Need the commit in the catch for this scenario.