Delayed showing tab until it's re-rendered
Go from tab A to tab B and change theme from light to dark, for example. Then switching back to A might show a flash of A with the light theme before the state update comes in shows the dark theme. That's because tab A was suspended when it wasn't active and the tab change happens on the native side.
Changed so that the tab change happens on the react side - the bottom tabs change on native side but delay the tab content from changing until the render happens. Also unfroze the selected tab instantly - it was incorrectly waiting until the useEffect fired. Dropped the !selected a while back (see commit ad1df29acd378db4b717e59bfd04451334b46139) when it was duplicate check but it isn't anymore so should be there.
Tested by blocking the [Home tab render](https://stackoverflow.com/a/16624104) in twitter sample and toggling the navigation barTint from blue to white. Switching tabs waits for the render and so there's no barTint flash