Delayed showing tab until it's re-rendered
Same change as applied to android. It 'prevents' the native change so there's no instant feedback - the android change is better because the selected tab changes immediately, it's just the content change that's delayed.
The shouldSelect is only called when the change is made native side. So return NO and send the change to React side and wait for the next render. The next render changes the selected tab and content together.