Added syncContent prop so user has control
When there's a theme change then set syncContent to true and then immediately set it to false. This marks all tabs as requiring a render before they're shown.

It's tricky because even if syncContent is now false, need to delay the render if the tab hasn't been shown since syncContent was true. So tracked a syncCounter that indicates if a tab is out of sync - once out of sync then it can only get back in sync after it's rendered and then shown.

There are 2 scenarios
1. Wait on theme change - increment the master counter whenever contentSync is toggled on
2. Always wait - increment the master counter whenever the tab becomes inactive and contentSync is on