Applied sync content changes to ios tabs
Same idea as on android. Block the tab press on the native side if the tab bar item is out of sync. The sync counter is held on the tab bar and each item has its own counter - if the counter is behind then send to js and wait for render.
The main counter is updated whenever contentSync toggled on (the theme change scenario) or whenever the tab is selected if contentSync is on (the always wait scenario).
Original userland plan was to do something like
onThemeChangePress={() => setContentSync(true)}
and reset it immediately
useEffect(() => {
  if (contentSync) setContentSync(false);
}, [contentSync]);
But when testing with weird blocking like below then had the reset happening before the set had actually been flushed to native. So the toggle didn't take effect.
while (currentTime + 2000 >= new Date().getTime()) {
}
New plan is to reset in onChangeTab instead
onChangeTab={() => setContentSync(false)}
This works - the reset happens in the next set state so doesn't make the exiting tab out of sync either