Renamed to from contentSync to preventFouc
Fouc stands for flash of unstyled content on the web so works well here. Don't want the name to be waitForRender or javascriptRender, for example, because should name it after the feature that it's implemented for. Toggling preventFouc from true to false will make tabs still wait until they're visited - not what you'd expect if it was called waitForRender, say. That's why better to name after the feature