Supported predictive back for material transitions
Thought it didn't work but it actually does. There are a lot of bugs that caused confused. For example, specifying only one transition throws weird errors using predictive back - it needs them all. Think it breaks with shared elements but will have another go.
But the key thing that was stopping it working was the VoidTransition. Weirdly, drag back then cancel was ok but on the second drag back the previous scene would disappear!? Only added the VoidTransition to turn off predictive back - think without it shared elements went funny (but will look again). So removed it and predictive back with material transitions works.
Still need to supply all transitions (enter, exit, reenter and return) but dummy in MaterialFade - has to be isSeekingSupported otherwise predictive doesn't work.
Also checked the fragment visibility before firing the onRest and onPopped because the onTransitionEnd fires when the predictive back is cancelled