Ensured swipe back works on api 33
With predictive back enabled in manifest but on api 33 (where predictive isn't supported) then, after swiping back, couldn't click on the previous scene.
Pretty sure there's (another) Android bug. The onBackStackChangeStarted was called differently on api 33 and 34.
Navigate A -> B -> C and swipe back then on Api 33 the crumb 2 (C) had isRemoving true (correct) and crumb 1 (B) has isRemoving false (true). But on Api 34 this is reversed - so crumb 2 isRemoving false and crumb 1 isRemoving true.
Had code in for the bug so it worked on Api 34 but broke on 33. Changed to cater for both scenarios. Used the keys.size instead lf the crumb to decide where going back to. Then doesn't matter which crumb isRemoving