Renamed stack context to fragment context
Used it from Dialog as well as NavigationStack. Need to created the stack nested in dialog in a whole different fragment manager - one not hosted by activity but custom hosted so can provide a different onBackPressDispatcher. Need to use the ComponentDailog's dispatcher so that the fragment manager can get the gesture back before the dialog gets it