Used showNow instead of custom code
Its the same code. Was only using the custom code because it was [recommended by android](https://developer.android.com/develop/ui/views/components/dialogs#FullscreenDialog). But couldn't use the android.R.id.content because it doesn't exist in react native and also the animation didn't work - not sure if that's because there's no content view? without the content view think it goes in separate window so transition doesn't work?!