Ran a stack inside modal (in twitter sample)
The onAfterUpdateTransaction of the NavigationStack was called before the one in DialogView. Didn't think this could happen - remember testing this when doing original predictive back PR - had tabs inside tabs inside tabs etc.
If the NavigationStack can't find the ancestor fragment then return early. It will get rerun later, when attached to window, for example.
Had to add the touch stuff to DialogRootView otherwise couldn't interact with the stack. Copied the code from the BottomSheetDialogView