Got predictive back working on dialog stack!!
Implemented a custom FragmentHostCallback so can get the FragmentManager to use the ComponentDialog's onBackPressedDispatcher. Without this, it uses the Activity's onBackPressedDispatcher - the Dialog's one always take precedence so the modal closes on the gesture back instead of moving back through the stack.

Had to do a lot of debugging to work out what to implement in the FragmentHostCallback.
Dispatching the start ensures the StackFragment is added (from the NavigationStack) - otherwise can't get a childFragmentManager from it.

Cloning the layout inflater in onGetLayoutInflater ensures that SceneFragment doesn't get the same layout inflater as the StackFragment - otherwise get an error saying factory already set on LayoutInflator. Got this clone inflator code from the FragmentActivity - the native FragmentHostCallback