Added the dialog fragment to the root frag manager
There's no need to put it in the ancestor stack. It doesn't use the fragment back stack to go back so doesn't need to be primary navigation. Clearer to put it in the top level support fragment manager. It starts a brand new support fragment manager of its own so makes sense to start a new ancestor stack with the dialog