Created a non-modal dialog - a sheet
This is like the [full screen recommendation from android](https://developer.android.com/develop/ui/views/components/dialogs#FullscreenDialog). Show the dialog fragment as a standard fragment. Can't add it to the android.R.id.content as per the recommendation because the content is too high up the view hierarchy for the fragment container to find it. Adding it to the support fragment manager would fix that but then the back gesture wouldn't work - the support fragment's handler is registered first so the main stack would take the gesture back instead of the support - so the dialog wouldn't close.
I like this compromise. Plus it seems in keeping with the non-modal bottom sheet - that doesn't take up the whole view neither.
Tried this before and couldn't get it to work - the trick was to add the dialog fragment to the same fragment manager - not the child. With the child got error saying already executing transaction when the modal stack tried to add fragment with commitAllowingStateLoss. Think it's ok to use the same fragment because the stack is actually A -> B -> Sheet A -> C - in this example, open the sheet and then click button to go to C in the scene behind the sheet (because it's non-modal).
Had to check the scene in onBackStackChangeStarted because it can now be the SheetViewFragment when the gesture closes the dialog.