Prevented error toggling modal from true to false
If don't stop then fragment state manager tries to call the getOnBackPressedDispatcher and onGetLayoutInflater - but there's no dialog so get error. There's a bug with the Dialog where the back listener isn't cleared but think that's fixed in later android.
Doesn't open the non-modal sheet but haven't investigated why yet