Ensured can navigate in stack behind non-modal
Navigate from A -> B then open modal and in background stack navigate from B -> C. Then gesture back needs to go back to B with modal open. It was going back and shutting the modal. That's because the NavigationStack back navigation was going to the back stack of crumb 1 (B) and so popping the modal off too.
Tagged the modal back stack with the same tag as the scene it's on - the crumb. It seems to work. Android takes the latest fragment with that tag (I think) and so keeps the modal open.
Can even do A -> B, open modal, navigate from D to E in modal, then navigate to C in stack behind. Then gesture back goes to B with modal on E, then gesture back goes to B with modal on D, then gesture back closes modal, then gesture back goes back to A