Started to move changes to bottom sheet
Couldn't get the expandedHeight to work on the modal dialog. Wasn't sure if it was because of the DialogRootView being a react native managed view instead of created in native. The height layoutparam always got reset to -1. So decided to use DialogRootView in the BottomSheetDialog and see if the expandedHeight still worked there. It does?! (Although the expanded offset isn't working?).
The back gesture also works on the bottom sheet in that it goes back through the stack - although the gesture back doesn't close the sheet? not sure why, but haven't tried bottom sheet gesture back before?!