Got the expanded offset and height back
It wasn't working when dialogrootview was created by react native. Not sure what the problem is but sometimes it would keep resizing until height went to 0 (with expanded offest). Put it back to how it used to be and seems ok.
There's a weirdness with scrolling with modal stack inside bottom sheet. Sometimes sheet will take the scroll away from content and sometimes not. Not sure