Implemented stack for non-modal bottom sheet
Copied the changes over from the non-modal sheet. Created a primary fragment to host the stack. Not 100% sure this is necessary for non-modals - they could probably use the parent stack fragment because modal navigation happens on top of current stack. But modal stacks need separate fragment managers - so keeps the code simpler to use the same approach in both cases. Also seems better to have modal stack in child fragment manager of parent stack.
The scrolling isn't working but think that's the custom nested scroll stuff in CoordinatorLayout - eeek!