Handled fragment controller lifecycle
Not sure how to really check this is right so went with dispatching and lifecycle as per the fragment methods. Rechecked the case that failed before (see commit 613f1ecd45285e4a28eb4a17fce35690eb475a8b).
Had to handle IllegalStateException in NavigationStackView because the fragment is already removed by dispatching the destroy.