Prevented coordinator intercepting sheet scrolls
Couldn't scroll in a bottomsheet or sheet because the coordinator was intercepting it. This interception is a workaround for [a (react native or android?) bug](https://github.com/facebook/react-native/pull/44099) where the nested scroll stops working of scroll content shorter than height.
Checked if the event target is a sheet and don't intercept in that case.