Enabled predictive back on non-modal bottom sheet
Makes it consistent with other sheets (like non-modal non-bottom). The default predictive back didn't work - on completion the sheet was hidden but then reappeared?! Tried requestLayout on CoordinatorLayoutView and that almost worked but still was a flicker. So forced it down the non-hideable path because that called setState instead of setStateInternal - this smoothed out the animation. Then set state to hidden if it was hideable.
Had to add onBackPressedCallback to the right dispatcher - activity if not inside dialog and componentdialog's if inside dialog. A -> B (C -> D) where C is dialog and D is scene in dialog with bottom sheet.