Removed gap before stack inside sheet is shown
Putting a modal inside a sheet (bottom sheet from before) didn't work - there was a gap before it showed. This is super weird because tried it in react native modal as there's no gap. Removing the animation (even though there isn't one for first scene) removed the gap - the stack showed straightaway. But no idea why it worked for react native modal where the animate flag was obviously the same - tried to find a difference in react native code but couldn't. Anyway, happy with the change because don't want to animated if 'first' scene (could be second scene if fluently navigating but still effectively the first if there aren't any previous.