Work for nested stacks inside nested modals
The stacks have different NavigationContext if there's stacks inside them - want to use the top level NavigationContext only because want to run the code when the main scene is active again. So used a SheetContext with onNavigated - it's only useNavigated for the root one and then the child ones are called iteratively from the top down. A scene/sheet can only open one sheet so ok to have single onNavigated.
Also turned off dismissViewControllerAnimated for now because this fails for similar reason. It's not just called for the top level scene/sheet anymore - opening a child sheet is also from a scene if it's
from a stack inside the modal. Plan to use the root on the SheetContext for this - only dismissViewControllerAnimated if the viewControllerToPresent is root (can cast to sheet and check)