Dismissed sheets only if root sheet
This prevents the problem of opening a child sheet with a modal stack closing all sheets. Had trouble remembering why this method was needed but then read commit 8f0f022b76c4967c199fe72670583f0b318c6e15 and finally remembered the timing issues. This method's needed when navigate from A with sheets open to B with a sheet open immediately. Without this dismiss with completion then the sheets on A will close but the sheet on B won't open