Ensured the right tab is shown
In twitter example, opened modal containing stack starting on notifications scene on fabric - not sure if this happened on non-fabric. About 1 out of 3 times the wrong tab would show - 'All' was active but mentions content was showing. Not sure why but setting the current item seems to fix it. Well, i did see it about 1 time out of 100. About 2 times out of 100 can see the other scene behind but can fix that by setting background color of the tab content