Prevented blank screen oddness when flinging sheet
Tried so hard to get view controller recycling to work (new fabric feature). But it just doesn't with view controllers. Example number 763 is when opening sheet B from sheet A then flinging sheet B to dismiss - very often this would give a shadow where sheet A should be (just the opaque background but no sheet A). Sometimes, if sheet A was still around, when closing sheet A would get an empty scene behind - so just the black underlay of the stack and no scene.
No clue where the problem is but preventing recycling fixes the issues. Was going to try and live with recycling because had it mostly working - maybe only errors sometimes on reload/refresh - but this was the final straw.
There's definitely some clash with recycling and uikit - something deep about uiviewcontrollers