Created Item view for non-primary tabs
On Fabric (not sure about old arch yet) the tab bar items sometimes overlap when opened in a Dialog. React Native's style overrides Androids style - instead of the hidden tab being offscreen, React Native renders it in same place as active tab. Creating dedicated item view for non-primary so they can needsCustomLayoutForChildren. Got away without this so far because it doesn't seem to happen outside of Modal (might've seen it once) - but really, if using Android's style, then should return true from needsCustomLayoutForChildren.
Hopefully won't need custom cpp for this even though need it for TabBarItemView - that's for images on ios, so plan to not use TabBarPagerItemView on ios - only needed on android. Try and keep TabBarItemView for non-primary on ios