Copied Item view for non-primary tabs from fabric
Tried the same scenario as caused the problem on fabric - sheet with notifications tab in twitter sample - but the problem doesn't happen. Guess the timing is different so the Android styling happens after the React Native styling. But creating the same dedicated non-primary tab bar item on old arch for consistency and correctness