Ensured drawer content is touchable
The NavigationView inserts its own NavigationMenuView child (the content is meant to be children of this) and its receiving the events instead of the actual content. This menu view is the last child so it gets the events. Could implement ReactZIndexedViewGroup to stop it receiving them - but simpler to exclude it from the tree. Think it's to give some material styling - can't support that because React Native does the styling. Can support material styling when there's no actual child content (like in BottomNavigationView) but not for the drawer where there's actual content