Removed custom navigation icon if auto
No point using a custom icon with drawer toggle because the whole point is the animation. Plus it got messy because creating the toggle again when navigating back (onAttached called again which created the toggle) stopped the animation from working. It replaced the navigation icon with the one from the previous toggle. Makes most sense to use the built in icons when auto navigate. If user wants custom icon then can manually navigate