Implemented auto navigation back on bottom bar
Copied over changes from toolbar/bottombar to searchbar. This is just for the home/back - will do drawer toggle too though. Had to do weird thing because couldn't get the back arrow to show - setting the navigationicon to null just didn't work with searchbar - when the button was added back it didn't have the right layout. It got added back in wrong place and took up the whole of the navigationbar. But if don't set it to null then setDisplayHomeAsUpEnabled won't set it to arrow. So overrode getNavigationIcon and returned null just while setting the homeAsUp - that way the ToolbarWidgetWrapper thinks it's null and sets it to arrow - and the Toolbar doesn't remove the view because never setNavigationIcon to null