Allowed custom navigation icon with auto
Improved the idea. Want a user to always turn on autoNavigation without having to think about whether it's first screen or not. Allowing a custom navigation icon is crucial for that as well as firing onNavigationPress (if it's not been taken over by auto navigation).
So always going to set the navigation icon if it's provided. If it's not provided then going to use the default one from auto navigation (or from search bar if auto isn't on).