Implemented auto navigation toggle on search bar
Copied over drawer toggle from bottombar to searchbar. Had to copy the DrawerArrowDrawable setup from material SearchView because there's no public property to set to add it - the useDrawerArrowDrawable is how to set it but that can only be set from resources styles. Was gonna try using resource but the code is only a few lines long so copied it in instead.
If don't add the drawerarrordrawable to the searchresults toolbar then android falls over because it goes down the FadeThroughDrawable path and DrawerArrowDrawable (on the searchbar) doesn't have constant state?!