Ensured navigation image tinted by default
The default navigation icon tint is null - at least on the DayNight.NoActionBar material3 theme. This is confusing for new users because it looks like the image isn't there. So defaulted to black if null. It's not null on bottombar and searchbar so leaving them.
Also switched to BlendModeColorFilter to avoid deprecation warning