Prevented error removing key path observer twice
This error happens on fabric when the search view is removed (tapping r to reload the zoom app gives the error). The unmountChild is called and this removes and then the dealloc is called which tries to remove and get the error. On the old arch for some reason the removeReactSubview isn't called (nor the dealloc).
Anyway, I rested the reason for the hidden keyPath observer on fabric (see commit 475f588f2998bdb6ce49ef514749f73f43b24368) and the problem isn't there on the new arch. Clicking the search bar doesn't show empty search results like it does on the old arch. So deleted the observer logic entirely.
Also added a safety check on old arch even though the double remove bug doesn't happen there