Sketched in idea for native style props on web
Idea is to use unmountStyle and crumbStyle props from native on the web. Original though was to change navigation motion to use these props - but new idea is to convert them into props that navigation motion understands. This makes more sense because navigation motion is like the 'primitive' so don't want to change that. The navigation native web layer turns props into native primitives so that feels like the right place to do it. Want navigation motion to stay independent of native