Implemented scene style overrides
Can't do unmountedStyle || () => {} like on the NavigationStack because mustn't override by default. Only override if there's an unmountStyle prop and also only if not doing custom render i.e. if using the new in-built render.
The override isn't quite right for duration override. But to get this right is much more complicated so skipping for now - complicated because need to implement mountedStyle, but what if only overriding unmountStyle and not crumbStyle - in that case mountedStyle has to call Scene's unmountStyle but NavigationStack's crumbStyle. Too messy to implement now, will wait and see if needed