Sketched in animation api idea
Check the flags and run the animations - either play or reverse depending on whether enter or exit. Set fill mode to forwards so that animation runs from where it left off. For example, if use arrows to flick between history back and forward the animation to play then reverse smoothly