Tidied up push/pop/enter/exit
Testing in the people/person master/details sample. Tried back/forward arrows when details came in and it didn't work. It wasn't correctly setting the nav state. It didn't think the details was pushEnter because it there was a scene before - it was mid popExit so the scene was there. It was only setting pushEnter when the scene wasn't there before.
Also tracked the current navigation state on the scene element because can't afford to play the animation twice. When removed the popExit scene it updated scenes and reevaluated animation - this replayed the same animation so it ran twice. So tracking to avoid this.