Added rest/freeze back
Not the same as before - not rendering on animation anymore so can't calculate rest from the scenes. Instead put it back to what it used to be and set it to true when the mounted scene rests. This matches native as well - don't need to wait for all scenes to rest. Only freeze crumb scenes anyway and if they're still animating then they're hidden anyway so should be fine. I guess if it's not a stack - some other custom navigation on mobile - then maybe still can see the crumbs?
Had to prevent onRest being called in infinite loop - so used the navState for that. Cleared it when they come to rest