Fixed animations when back twice
In twitter example only the 1st back worked. After that nothing happened. Needed to persist the animation otherwise the original pushEnter animation was removed when it was replaced by the pushExit (wrongly named - it should be popEnter).
Also only cleared navState if the animation ran - the finished promise always resolves if no animation so navState was being cleared when actually different animation was running