Fixed unit tests
A bit early to be fixing unit tests but wanted to check the approach is sound. Added id={key} to scenes to get the tests passing - will think what to really do later. Excluded the exiting scenes immediately because not animating in unit tests. Had to fire onRest otherwise tests failed where was expecting the crumb scene to be suspended (no state update).
Tweaked a couple of tests so that the useLayoutEffect fired - separated into separate acts. Think the old way of running requestAnimationFrame in a row meant it wasn't necessary. Deleted all the requestAnimationFrame because that's not there anymore.
Might end up keeping old NavigationMotion and creating NavigationStack for backward compatibility - so might need to bring back the old test approach