Supported keeping history back/forward pressed
This is continuous overlapping navigations. For example,  A -> B -> C -> D then keeping finger on CMD + back arrow and scene B needs to do popEnter and popExit - so want the popEnter animation to complete and then the pushEnter reverse to run. So moved pushEnter logic inside the popExit finished promise - and vice versa. So pushEnter waits on popEnter and vice versa.
This was something couldn't get working properly with the old approach - because B would go from crumb to exit in a single duration. The old Motion would just move between target and destination so had no notion of intermediate states