Fixed back/forward during a replace navigation
Navigate from A -> B to A -> C then during the animation go browser back and forward. This showed a new scene animating in instead of reversing C back out. That's because it was given a different key - it had key 1+ because of replace but then going back and forward gave C key of 1 - because not replacing this time.
So changed the way the keys are built - used the state key names instead - a scene's key is all the crumb state's and current state's keys. So key looks like A-C for scene C both times. Then it works.
Have to smarten up the keys though because there could be state called A and A- and C and -C. Then could get same key names A--C for 2 different paths