Ensured replacing scene above replaced
Navigation A -> B to A -> C then replacing C should appear above replaced scene B. Used to use the key length to do the sorting but the key is A->B not 1+ anymore so can't use length. Stamped scenes with count so that later created ones always have higher count. Need scenes to keep the count they're stamped with so that toggling browser back and forward during the replacing animation keeps the fixed order with replacing on top - changing the count would see the order change