Fixed when animating multiple history back at once
A -> B -> C then go back 2 (used browser history drop down but can to navigateBack(2)) and while that's animating go forward once in history (back to B). This used to go blank, just black background. That's because B didn't go through the popEnter - it went straight from pushExit to popExit so couldn't play pushEnter because  it hadn't been reversed yet. So changed to take it through popEnter first. It will then go to popExit when the first popExit completes because then it rerenders