Fixed failing tests after previous change
Previous change means A -> B -> C and then navigatingBack 2 needs multiple renders to get rid of B. Now B first goes to popEnter and then popExit. So 2 renders when before it went straight to popExit. Had to change because going straight from pushExit to popExit skipped animation so went blank when forward in history during the animation.
So fixed the test mock animation to call onRest for popEnter to fire that extra render