Prevented errors when testing
Unregistering shared element threw because scene is null. Not sure need to unregister now because the animation only runs for fixed duration?
The shared elements kept rendering because when reversing used shared element instead of next shared element. Was only using shared to keep the same html element so just spread that on individually