Paused transition until shared elements ready
Now using web animations api there's no state update to trigger a rerender to get the shared elements. So paused the transition until the shared elements ready. When they're ready triggered a rerender to unpause and start shared elements - this is just the rough pause idea but tested with twitter sample and it does pick up the shared elements (before the pause it didn't - well not unti the transition finished). Hard-code the pause but will need to only pause if there are shared elements to check - maybe use sharedElements prop on NavigationStack like android (or maybe just check how many shared elements are in the from scene - but this isn't great because there might be different ones depending on the navigation)