Implemented proper pause
Checked for shared elements - copied approach from native stack. Call the sharedElements prop (new on mobile) to see if there are any shared els. If they're not all in the registry then need to pause. Stored the crumb length in the pause because plan to use that for the timeout mechanism (not yet implemented though)