Implemented pause timeout
Can't pause forever waiting for the shared elements. So added timeout like there is on native android so it just transitions if the target shared element doesn't render before 300ms.
Also unregistered shared elements on navigation - Navigation A -> B then click to go to C from A while scene B animates in. If the same shared element is on C and B then it was picking up the one from B because B is still animating so it's still there - then when the shared element actually started B had been removed so it couldn't find the closest scene. It picked up the wrong shared element (the one from the leaving scene)