Fixed shared elements for interruptions
Backed out the previous 'fix' that unregisterSharedElement because that broke basic shared elements on going back - they weren't there anymore.
So this fixes the problem of A -> B and while B is animating click on A to C with a shared element. B would clear out the shared elements so they weren't there for C. So changed the shared elements to cache by url instead of key - years ago they used to be by url, but the comment didn't explain why they changed. I guess so they were easier to clear - but not even sure clearing is necessary from stack because they clear themselves.
Even after changing to urls it still didn't work - turns out B still cleared out shared elements. That's because when animating A-> B then clicking A to C it changed scene B to C! Never even noticed this before. It even happens with NavigationMotion. So changed Scene to use the url instead of the crumbs as well - so it only updates its navigationEvent if the url matches - that way it will stick with B when C is coming in