Hid child shared elements of a shared element
If there's nested shared elements then don't want the parent to show the child because the child will be shown separately. So marked shared elements with data attribute and hid them during the transition. These are only the dynamically created elements so no need to show them again - they only live for the transition