Fixed shared el motion on current twitter
Tested current twitter for regressions and it fell over because it was filtering the shared elements by data. It was trying to prevent sharing photo when going from tweets to tweets.
Had changed the way the data is held on the shared elements so it wasn't available. Put back the way shared elements returned from the registry (even though they're held differently - they're held on the element to prevent infinite render when data updates) so that backward compatible.
Do that in a different way now - added the sharedElements prop to stack to match android. So only share if specifically list the shared elements