Sketched in removal of unmounted scenes
Removed unmounted whenever there's a non-backward or non-history navigation. It shouldn't really matter if they're not removed - for example, if they're still animating when they're due to be removed. Should be ok to keep them around because they won't be picked up if sort of the keys - plan is to increment the keys when there's a forward navigation (using counter idea from native) - they'll be removed eventually