Prevented flicker when navigating forward
The pop exit animation wasn't being retained. After exiting it was actually appearing over the top of the current scene - it was suspended so didn't display. Then when navigating forward it flash visible on top before starting the push enter animation. Changed fill mode so that it stays offscreen when it's unmounted (in it's unmountStyle initial keyframe)